LPS-44643 Move closer to BreadcrumbEntry
